Split sample app into 2 separated demos
