+++ app/ambient_encounter_demo/ambient_encounter_demo.html
@@ -1,0 +1,167 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Nabla Ambient Encounter Demo</title>
+    <link rel="stylesheet" href="../shared/sharedStyle.css">
+    <link rel="stylesheet" href="style.css">
+    <script type="module" src="ambient_encounter_demo.js"></script>
+</head>
+<body>
+    <div class="container">
+        <h1>Nabla Ambient Encounter Demo</h1>
+        <a href="../index.html" class="back-link">← Back to demo selection</a>
+
+        <hr>
+
+        <div id="token-error" class="hide"></div>
+
+        <div class="filters">
+            <label for="first-transcript-locale">First transcript locale:</label>
+            <select id="first-transcript-locale">
+                <option value="ENGLISH_US" selected="selected">ENGLISH_US</option>
+                <option value="ENGLISH_UK">ENGLISH_UK</option>
+                <option value="FRENCH_FR">FRENCH_FR</option>
+                <option value="SPANISH_ES">SPANISH_ES</option>
+                <option value="SPANISH_MX">SPANISH_MX</option>
+                <option value="ARMENIAN_AM">ARMENIAN_AM</option>
+                <option value="BENGALI_IN">BENGALI_IN</option>
+                <option value="CANTONESE_CN">CANTONESE_CN</option>
+                <option value="CROATIAN_HR">CROATIAN_HR</option>
+                <option value="FILIPINO_PH">FILIPINO_PH</option>
+                <option value="GERMAN_DE">GERMAN_DE</option>
+                <option value="GREEK_GR">GREEK_GR</option>
+                <option value="GUJARATI_IN">GUJARATI_IN</option>
+                <option value="HINDI_IN">HINDI_IN</option>
+                <option value="ITALIAN_IT">ITALIAN_IT</option>
+                <option value="JAPANESE_JP">JAPANESE_JP</option>
+                <option value="KHMER_KH">KHMER_KH</option>
+                <option value="KOREAN_KR">KOREAN_KR</option>
+                <option value="MANDARIN_CN">MANDARIN_CN</option>
+                <option value="POLISH_PL">POLISH_PL</option>
+                <option value="PORTUGUESE_PT">PORTUGUESE_PT</option>
+                <option value="PUNJABI_IN">PUNJABI_IN</option>
+                <option value="RUSSIAN_RU">RUSSIAN_RU</option>
+                <option value="SERBIAN_RS">SERBIAN_RS</option>
+                <option value="TAMIL_IN">TAMIL_IN</option>
+                <option value="TELUGU_IN">TELUGU_IN</option>
+                <option value="THAI_TH">THAI_TH</option>
+                <option value="VIETNAMESE_VN">VIETNAMESE_VN</option>
+                <option value="ARABIC_LB">ARABIC_LB</option>
+                <option value="ARABIC_EG">ARABIC_EG</option>
+                <option value="ARABIC_SA">ARABIC_SA</option>
+                <option value="ARABIC_MA">ARABIC_MA</option>
+                <option value="HEBREW_IL">HEBREW_IL</option>
+                <option value="URDU_IN">URDU_IN</option>
+                <option value="PERSIAN_IR">PERSIAN_IR</option>
+            </select>
+        </div>
+
+        <div class="filters">
+            <label for="second-transcript-locale">Second transcript locale:</label>
+            <select id="second-transcript-locale">
+                <option value="ENGLISH_US">ENGLISH_US</option>
+                <option value="ENGLISH_UK">ENGLISH_UK</option>
+                <option value="FRENCH_FR">FRENCH_FR</option>
+                <option value="SPANISH_ES" selected="selected">SPANISH_ES</option>
+                <option value="SPANISH_MX">SPANISH_MX</option>
+                <option value="ARMENIAN_AM">ARMENIAN_AM</option>
+                <option value="BENGALI_IN">BENGALI_IN</option>
+                <option value="CANTONESE_CN">CANTONESE_CN</option>
+                <option value="CROATIAN_HR">CROATIAN_HR</option>
+                <option value="FILIPINO_PH">FILIPINO_PH</option>
+                <option value="GERMAN_DE">GERMAN_DE</option>
+                <option value="GREEK_GR">GREEK_GR</option>
+                <option value="GUJARATI_IN">GUJARATI_IN</option>
+                <option value="HINDI_IN">HINDI_IN</option>
+                <option value="ITALIAN_IT">ITALIAN_IT</option>
+                <option value="JAPANESE_JP">JAPANESE_JP</option>
+                <option value="KHMER_KH">KHMER_KH</option>
+                <option value="KOREAN_KR">KOREAN_KR</option>
+                <option value="MANDARIN_CN">MANDARIN_CN</option>
+                <option value="POLISH_PL">POLISH_PL</option>
+                <option value="PORTUGUESE_PT">PORTUGUESE_PT</option>
+                <option value="PUNJABI_IN">PUNJABI_IN</option>
+                <option value="RUSSIAN_RU">RUSSIAN_RU</option>
+                <option value="SERBIAN_RS">SERBIAN_RS</option>
+                <option value="TAMIL_IN">TAMIL_IN</option>
+                <option value="TELUGU_IN">TELUGU_IN</option>
+                <option value="THAI_TH">THAI_TH</option>
+                <option value="VIETNAMESE_VN">VIETNAMESE_VN</option>
+                <option value="ARABIC_LB">ARABIC_LB</option>
+                <option value="ARABIC_EG">ARABIC_EG</option>
+                <option value="ARABIC_SA">ARABIC_SA</option>
+                <option value="ARABIC_MA">ARABIC_MA</option>
+                <option value="HEBREW_IL">HEBREW_IL</option>
+                <option value="URDU_IN">URDU_IN</option>
+                <option value="PERSIAN_IR">PERSIAN_IR</option>
+            </select>
+        </div>
+
+        <div class="filters">
+            <label for="note-locale">Note locale:</label>
+            <select id="note-locale">
+                <option value="ENGLISH_US">ENGLISH_US</option>
+                <option value="ENGLISH_UK">ENGLISH_UK</option>
+                <option value="FRENCH_FR">FRENCH_FR</option>
+            </select>
+        </div>
+
+        <div class="filters">
+            <label for="note-template">Note template:</label>
+            <select id="note-template">
+                <option value="GENERIC_MULTIPLE_SECTIONS">General Medicine</option>
+                <option value="GENERIC_SOAP">SOAP</option>
+            </select>
+        </div>
+
+        <div class="filters">
+            <label for="patient-context">Patient context:</label>
+            <textarea id="patient-context"></textarea>
+        </div>
+
+        <div class="filters">
+            <label for="note-sections">Section to customize below:</label>
+            <select id="note-sections"></select>
+        </div>
+
+        <div id="section-customization-fields">
+            <div class="filters">
+                <label for="style-select">Style:</label>
+                <select id="style-select">
+                    <option value="AUTO">Auto</option>
+                    <option value="PARAGRAPH">Paragraph</option>
+                    <option value="BULLET_POINTS">Bullet points</option>
+                </select>
+            </div>
+
+            <div class="filters">
+                <label for="custom-instruction">Custom instruction:</label>
+                <textarea id="custom-instruction" placeholder="Your custom instruction"></textarea>
+            </div>
+        </div>
+
+        <hr>
+
+        <div>
+            <button id="start-btn">Start consultation</button>
+            <button id="generate-btn" disabled>Generate note</button>
+            <button id="normalize-btn">Generate normalized data</button>
+            <button id="patient-instructions-btn">Generate patient instructions</button>
+            <button id="clear-btn">Clear</button>
+        </div>
+
+        <div id="transcript" class="results">
+            <h3>Transcript:</h3>
+        </div>
+        <div id="note" class="results">
+            <h3>Note:</h3>
+        </div>
+        <div id="normalized-data" class="results">
+            <h3>Normalized data:</h3>
+        </div>
+        <div id="patient-instructions" class="results">
+            <h3>Patient instructions:</h3>
+        </div>
+    </div>
+</body>
+</html>
