+++ app/dictated_note_demo/dictated_note_demo.html
@@ -1,0 +1,79 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Nabla Dictated Note Demo</title>
+    <link rel="stylesheet" href="../shared/sharedStyle.css">
+    <link rel="stylesheet" href="style.css">
+    <script type="module" src="dictated_note_demo.js"></script>
+</head>
+<body>
+    <div class="container">
+        <h1>Nabla Dictated Note Demo</h1>
+        <a href="../index.html" class="back-link">← Back to demo selection</a>
+
+        <hr>
+
+        <div id="token-error" class="hide"></div>
+
+        <div class="filters">
+            <label for="dictationLocale">Dictation locale:</label>
+            <select id="dictationLocale">
+                <option value="ENGLISH_US" selected="selected">ENGLISH_US</option>
+                <option value="ENGLISH_UK">ENGLISH_UK</option>
+                <option value="FRENCH_FR">FRENCH_FR</option>
+                <option value="SPANISH_ES">SPANISH_ES</option>
+                <option value="SPANISH_MX">SPANISH_MX</option>
+                <option value="ARMENIAN_AM">ARMENIAN_AM</option>
+                <option value="BENGALI_IN">BENGALI_IN</option>
+                <option value="CANTONESE_CN">CANTONESE_CN</option>
+                <option value="CROATIAN_HR">CROATIAN_HR</option>
+                <option value="FILIPINO_PH">FILIPINO_PH</option>
+                <option value="GERMAN_DE">GERMAN_DE</option>
+                <option value="GREEK_GR">GREEK_GR</option>
+                <option value="GUJARATI_IN">GUJARATI_IN</option>
+                <option value="HINDI_IN">HINDI_IN</option>
+                <option value="ITALIAN_IT">ITALIAN_IT</option>
+                <option value="JAPANESE_JP">JAPANESE_JP</option>
+                <option value="KHMER_KH">KHMER_KH</option>
+                <option value="KOREAN_KR">KOREAN_KR</option>
+                <option value="MANDARIN_CN">MANDARIN_CN</option>
+                <option value="POLISH_PL">POLISH_PL</option>
+                <option value="PORTUGUESE_PT">PORTUGUESE_PT</option>
+                <option value="PUNJABI_IN">PUNJABI_IN</option>
+                <option value="RUSSIAN_RU">RUSSIAN_RU</option>
+                <option value="SERBIAN_RS">SERBIAN_RS</option>
+                <option value="TAMIL_IN">TAMIL_IN</option>
+                <option value="TELUGU_IN">TELUGU_IN</option>
+                <option value="THAI_TH">THAI_TH</option>
+                <option value="VIETNAMESE_VN">VIETNAMESE_VN</option>
+                <option value="ARABIC_LB">ARABIC_LB</option>
+                <option value="ARABIC_EG">ARABIC_EG</option>
+                <option value="ARABIC_SA">ARABIC_SA</option>
+                <option value="ARABIC_MA">ARABIC_MA</option>
+                <option value="HEBREW_IL">HEBREW_IL</option>
+                <option value="URDU_IN">URDU_IN</option>
+                <option value="PERSIAN_IR">PERSIAN_IR</option>
+            </select>
+        </div>
+
+        <div class="filters">
+            <span style="margin-right: 8px;">Explicit punctuation:</span>
+            <label class="switch">
+                <input id="punctuation-switch" type="checkbox" checked>
+                <span class="slider round"></span>
+            </label>
+        </div>
+
+        <hr>
+
+        <div>
+            <button id="dictate-btn">Start</button>
+            <button id="pause-btn" disabled>Pause</button>
+        </div>
+
+        <div id="dictated-note" class="results">
+            <h3>Dictated note:</h3>
+        </div>
+    </div>
+</body>
+</html>
