+++ app/shared/sharedStyle.css
@@ -1,3 +1,5 @@
+/* Common styles for Nabla API demos */
+
 body {
     display: flex;
     justify-content: center;
@@ -10,6 +12,94 @@ body {
     width: 960px;
     overflow-y: auto;
 }
+
+/* Demo selector specific styles */
+
+.demo-grid {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 20px;
+    margin-top: 20px;
+    justify-content: center;
+}
+
+.demo-card {
+    flex: 0 0 calc(50% - 20px);
+    max-width: 340px;
+    min-width: 280px;
+    border: 1px solid #ddd;
+    border-radius: 8px;
+    padding: 25px 20px;
+    text-align: center;
+    box-shadow: 0 2px 5px rgba(0,0,0,0.05);
+    transition: all 0.3s ease;
+    background-color: #fff;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    height: auto;
+}
+
+.demo-card:hover {
+    transform: translateY(-5px);
+    box-shadow: 0 8px 15px rgba(0,0,0,0.1);
+    border-color: #2196F3;
+}
+
+.demo-card h2 {
+    margin-top: 0;
+    color: #333;
+    font-size: 20px;
+    margin-bottom: 15px;
+}
+
+.demo-card p {
+    color: #666;
+    margin-bottom: 25px;
+    font-size: 14px;
+    line-height: 1.5;
+    flex-grow: 1;
+}
+
+.demo-btn {
+    display: inline-block;
+    background-color: #2196F3;
+    color: white !important;
+    padding: 10px 20px;
+    text-decoration: none;
+    border-radius: 4px;
+    font-weight: bold;
+    transition: background-color 0.2s;
+    box-shadow: 0 2px 5px rgba(33, 150, 243, 0.3);
+}
+
+.demo-btn:hover {
+    background-color: #0b7dda;
+    box-shadow: 0 4px 8px rgba(33, 150, 243, 0.5);
+}
+
+.note-section {
+    margin-top: 40px;
+    padding: 15px;
+    background-color: #e8f4fe;
+    border-left: 4px solid #2196F3;
+    border-radius: 4px;
+}
+
+/* For smaller screens */
+@media (max-width: 768px) {
+    .demo-card {
+        flex: 0 0 100%;
+        max-width: 100%;
+    }
+    
+    .container {
+        padding: 15px;
+        margin: 10px;
+    }
+}
+
+/* Common elements */
 
 .results {
     border: 1px black dashed;
